requirements.txt
fastapi
uvicorn
langchain
langchain-groq
langchain-core
langchain-community
langgraph
python-dotenv
pydantic
openpyxl
pdfplumber
python-docx
chromadb